add double

+++ include/lexema.h
@@ -6,7 +6,7 @@ enum TypeLex { number, binOp, unOp, leftB, rightB };
 
 struct Lexema {
 	TypeLex type;
-	long long value;
+	double value;
 	std::string name;
 	int priority;
 };
